Split functions and refactor much of the code
