+++ src/manifest.json
@@ -1,0 +1,63 @@
+{
+    "name": "Skyward GPA Calculator - And More",
+    "description": "This is an extension which injects various scripts into Skyward to calculate GPA and perform other tasks.",
+    "version": "0.13",
+    "permissions": [
+        "tabs",
+        "scripting",
+	"storage"
+    ],
+    "host_permissions": [
+        "http://*/scripts/wsisa.dll/WService=ws*/*",
+	"https://*/scripts/wsisa.dll/WService=ws*/*"
+    ],
+    "icons": {                   
+            "16": "../assets/icons/icon16.png",
+            "38": "../assets/icons/icon38.png",
+            "48": "../assets/icons/icon48.png",
+            "128": "../assets/icons/icon128.png"  
+    },    
+    "action": {
+    "default_title": "Skyward GPA Calculator",
+    "default_popup": "index.html",
+            "default_icon": {                   
+                    "16": "../assets/icons/icon16.png",
+                    "38": "../assets/icons/icon38.png",
+                    "48": "../assets/icons/icon48.png",
+                    "128": "../assets/icons/icon128.png"  
+            }
+    },
+    "web_accessible_resources": [{
+        "resources": [
+            "../css/variables.css",
+            "../css/qsfmain001.css",
+            "../css/sfhome001.css",
+            "../css/jquery.cluetip.css",
+            "../css/sfgradebook001.css",
+            "../css/qclssbase001.css",
+            "../css/qclsslogin001.css",
+            "../sprite/arrows_sprite.png",
+            "../sprite/printer1.png",
+            "../sprite/sf_bg_sprite001.png",
+            "../sprite/sf_icons_sprite.png",
+            "../gpachart.html",
+            "../assets/images/SkyLogoBlue.png",
+            "script2.js"
+        ],
+        "matches": ["https://skyward.iscorp.com/*", "http://skyward.iscorp.com/*"]
+    }],
+    "content_scripts": [{
+        "matches": [
+            "http://*/scripts/wsisa.dll/WService=*",
+	    "https://*/scripts/wsisa.dll/WService=*"
+        ],
+        "js": ["main.ts"],
+        "run_at": "document_end"
+    }],
+    "applications": {
+        "gecko": {
+            "id": "YOUR_ADDON_ID"
+        }
+    },
+    "manifest_version": 3
+}
