+++ src/settings.ts
@@ -1,0 +1,6 @@
+export interface Settings
+{
+    autosaveSetting: boolean
+    originalGraphHasSet: number,
+    graphHasSet: number,
+}
